requirements.txt
asgiref==3.9.1
channels==4.0.0
channels-redis==4.2.0
Django==5.2.6
django-environ==0.12.0
django-js-asset==3.1.2
django-mptt==0.18.0
django-redis==5.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.10.1
pyotp==2.9.0
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.24.0
websockets==12.0

django-filter==24.3
firebase-admin==6.5.0  # Firebase Cloud Messaging
django-debug-toolbar==4.4.6
django-cors-headers==4.3.1